requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
passlib
bcrypt==4.0.1
python-jose
python-multipart